pyproject.toml
[tool.poetry]
name = "duck-flow"
version = "0.1.0"
description = ""
authors = ["yadavi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
duckdb = "^1.3.2"
google-cloud-bigquery = "^3.34.0"
google-auth = "^2.40.3"
google-cloud-bigquery-storage = "^2.32.0"
pyarrow = "^20.0.0"
pandas = "^2.3.1"
fire = "^0.7.0"
loguru = "^0.7.3"
pydantic = "^2.11.7"
pytest = "^8.4.1"
ruff = "^0.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
